+++ src/index.ts
@@ -5,6 +5,7 @@ export { SyntaxParser as SyntaxEncode } from "./v1/compiler/syntax-encode.js";
 export { LexerParser } from "./v1/compiler/lexema-encode.js";
 export type { Token } from "./v1/compiler/token.js";
 export { DON, Directive, HeredocValue } from "./don.js";
+export type { Loc } from "./don.js";
 export {
   DirectiveJSONEncoder,
   DirectiveJSONDecoder,